Bug 19028: Add a note to the pref description
The idea was to keep the current behaviour, plus adding the option to
display on a separate column.
We need an explanation on the syspref description, stating that it is not enough to
choose 'on a separate column', enabling the colum in the columns
settings is required for it to take effect.

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>