Bug 19584: Check compare barcodes box when uploading barcodes
This patch makes two trivial changes in the inventory interface:

[1] If you have uploaded a barcode file, the compare barcodes checkbox is
automatically checked. In most cases this will be the option the user wants.
The compare will report missed items on the shelf.
[2] If the compare barcodes checkbox is checked, the Skip items on loan-
checkbox is automatically checked. This again is often desirable. If you
compare, you normally are not interested in seeing items reported missing
while they are checked out.

Test plan:
[1] Upload a barcode file on inventory. Is the compare checked?
[2] Toggle compare. Does 'skip items on loan' change accordingly?

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>

Signed-off-by: Grace Smyth <gracesmythh@gmail.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>