Bug 19744: Move template JavaScript to the footer: Offline circulation
This patch modifies the staff client offline circulation templates so
that JavaScript is included in the footer instead of the header.

To test, apply the patch and test the JavaScript-driven features of the
modified templates: All button controls, DataTables functionality, tabs,
etc.

Circulation -> Built-in offline circulation interface
  -> Synchronize
   -> Download records
  -> Check out
  -> Check in
    -> Synchronize
      -> Upload transactions
  -> Pending offline circulations
    -> Check all, uncheck all
    -> Delete
    -> Process

Circulation -> Upload offline circulation data
  -> Upload .koc file (gerated by the Koha Offline Circulation program,
     for instance).

Signed-off-by: Simon Pouchol <simon.pouchol@biblibre.com>

Signed-off-by: Josef Moravec <josef.moravec@gmail.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>