Bug 19786: Move template JavaScript to the footer: Authorities, part 2
This patch modifies the staff client patron lists templates so that
JavaScript is included in the footer instead of the header.

To test, apply the patch and test the JavaScript-driven features of
each modified template.

I've made one change to the JavaScript in addition to moving it: I've
made it so that the blank window which pops up briefly in this process
is 100px x 100px instead of full screen.

- Cataloging -> Add or edit bibliographic record in a framework which
  has authorities linked to a tag
  -> Click authorities plugin link
     -> Create new authority button
     -> Autocomplete on text inputs (except "Search all headings")
     -> Search
        -> Select authority record ("choose")
           -> Click authorities plugin link again
              -> Clear field

Signed-off-by: Claire Gravely <claire_gravely@hotmail.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>