Bug 19790: Remove additionalauthors.author from installer files
In the upgrade from 2.2 to 3.0 (some time ago already) the table
additionalauthors has been dropped. Unfortunately, a reference to this
table has not been removed from the MARC framework installer files.

This patch clears kohafield for 700$a or 200$g in 32 installer files.

Note: The reference to additionalauthors in field kohafield is silently
ignored in AddBiblio/ModBiblio since the routines _koha_add_biblio,
_koha_add_biblioitem, _koha_modify_biblio and
_koha_modify_biblioitem_nonmarc simply do not use it. You can define a
Koha to MARC mapping for a biblio/biblioitem field but if the field is not
referenced in one of those routines, nothing happens..
Also note that C4::Items::_koha_modify_item does not hardcode all fields
and will respond to a new item mapping.

Test plan:
[1] Run a new install or verify somehow that the change to these sql files
    is correct.
    The change is a result of: sed -i -e "s/'additionalauthors.author'/''/g"

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Run a new install with MARC21 and English. This includes the mandatory
file marc21_framework_DEFAULT.sql. Verified that 700$a was inserted
correctly (without additionalauthors).

Signed-off-by: Charles Farmer <charles.farmer@inLibro.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>