Bug 19823: Move template JavaScript to the footer: MARC21 editor plugins
This patch modifies the templates for MARC21 editor plugins so that
JavaScript is included in the footer.

To test, you must link the plugins to their corresponding tag +
subfield (if they aren't already):

In a bibliographic framework:

000 -> marc21_leader.pl
006 -> marc21_field_006.pl
007 -> marc21_field_007.pl
008 -> marc21_field_008.pl
773$t -> marc21_linking_section.tt

In an authorities framework:

000 -> marc21_leader_authorities.tt
008 -> marc21_field_008_authorities.tt
    -> marc21_field_008_classifications.tt

For each field, trigger the plugin by clicking the plugin link. Test
JavaScript-driven functionality in each plugin window and confirm that
your selections are correctly saved to the editor.

Signed-off-by: Claire Gravely <claire.gravely@bsz-bw.de>

Edit: Fix for QA, add KOHA_VERSION variables to js includes

Signed-off-by: Josef Moravec <josef.moravec@gmail.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>