Bug 19869: Move template JavaScript to the footer: UNIMARC editor plugins, part 3
This patch updates even more UNIMARC plugin templates so that
JavaScript is included in the footer following the same pattern as other
templates which have been modified to fix Bug 17858.

To test you do not need a UNIMARC system.

Apply the patch and for each plugin, configure a MARC subfield (e.g.
100$a) to use that plugin.

From the MARC edit page, trigger the plugin and confirm that changes
made in the popup window are saved to the corresponding field in the
editor.

- unimarc_field_123a.tt
- unimarc_field_123d.tt
- unimarc_field_123e.tt
- unimarc_field_123f.tt
  - This template has been modified, but doesn't appear to be used. The
    unimarc_field_1234.pl script uses unimarc_field_123g.tt as its
    template.
- unimarc_field_123g.tt
- unimarc_field_123i_j.tt

Signed-off-by: Anne-Claire Bernaudin <anne-claire.bernaudin@univ-rennes1.fr>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>