Bug 19623: (follow-up) Correct footer include in pop-up windows
This patch corrects the footer include in two templates for Z39.50
pop-up windows. popup-bottom.inc shouldn't be used because it doesn't
have the JavaScript embed code in it.

To test, perform a Z39.50 search from both the MARC bibliographic editor
and the MARC authority editor. In both cases the Z39.50 popup window
should work correctly.

Signed-off-by: Josef Moravec <josef.moravec@gmail.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>