Bug 19877: Move template JavaScript to the footer: UNIMARC editor plugins, part 6
This patch updates more UNIMARC plugin templates than ever before, so
that JavaScript is included in the footer following the same pattern as
other templates which have been modified to fix Bug 17858.

To test you do not need a UNIMARC system.

Apply the patch and for each plugin, configure a MARC subfield (e.g.
100$a) to use that plugin.

From the MARC edit page, trigger the plugin and confirm that changes
made in the popup window are saved to the corresponding field in the
editor.

- unimarc_field_128a.tt
- unimarc_field_128b.tt
- unimarc_field_128c.tt
- unimarc_field_130.tt
- unimarc_field_135a.tt
- unimarc_field_140.tt
- unimarc_field_141.tt

Signed-off-by: delaye <stephane.delaye@biblibre.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>