Bug 19878: Move template JavaScript to the footer: UNIMARC editor plugins, part 7
This patch updates the last of the UNIMARC plugin templates so that
JavaScript is included in the footer following the same pattern as
other templates which have been modified to fix Bug 17858.

To test you do not need a UNIMARC system.

Apply the patch and for each plugin, configure a MARC subfield (e.g.
100$a) to use that plugin.

From the MARC edit page, trigger the plugin and confirm that changes
made in the popup window are saved to the corresponding field in the
editor.

- unimarc_field_210c.tt
- unimarc_field_210c_bis.tt
- unimarc_field_225a.tt
- unimarc_field_225a_bis.tt
- unimarc_field_4XX.tt
- unimarc_field_700-4.tt
- unimarc_leader.tt

Signed-off-by: Anne-Claire Bernaudin <anne-claire.bernaudin@univ-rennes1.fr>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>