Bug 19947: Update popup window templates to use Bootstrap grid: Acquisitions transfer order
This patch updates the acquisitions transfer order template to use the
Bootstrap grid instead of the YUI grid.

Also fixed: Corrected footer include for popup window.

To test, apply the patch and go to Acquisitions -> Vendor -> Open order
-> Transfer.

In the popup window search for a vendor, select, and choose a basket.

Signed-off-by: Charlotte Cordwell <charlotte.cordwell123@gmail.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>