Bug 19952: Update popup window templates to use Bootstrap grid: UNIMARC cataloging plugins
This patch updates two UNIMARC cataloging plugin templates to use the
Bootstrap grid instead of the YUI grid.

- unimarc_field_210c_bis.tt
- unimarc_field_686a.tt

To test you do not need a UNIMARC system.

Apply the patch and for each plugin, configure a MARC subfield (e.g.
100$a) to use that plugin.

From the MARC edit page, trigger the plugin and confirm that the
page in the popup window looks correct. Confirm that changes made in the
popup window are saved to the corresponding field in the editor.

Signed-off-by: Te Rauhina Jackson <terauhina.jackson@gmail.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>